refactor:update FlyoutItems component;feat:add HomeClient component;fix:remove comments
