+++ src/app/page.tsx
@@ -1,18 +1,5 @@
-import Page from '../components/Page';
-
-export async function fetchInvoicesPages(): Promise<number> {
-  const response = await fetch(`https://swapi.dev/api/people`, {
-    cache: 'no-store',
-  });
-  const total = await response.json();
-  return Math.ceil(total.count / 10);
-}
-export async function getPeople(page: number) {
-  const response = await fetch(`https://swapi.dev/api/people?page=${page}`, {
-    cache: 'no-store',
-  });
-  return response.json();
-}
+import { fetchInvoicesPages, getPeople } from '../redux/services/api_people';
+import HomeClient from '../components/HomeClient/HomeClient';
 
 export default async function Home({
   searchParams,
@@ -30,6 +17,10 @@ export default async function Home({
   const { results } = data;
 
   return (
-    <Page data={results} totalPages={totalPages} currentPage={currentPage} />
+    <HomeClient
+      data={results}
+      totalPages={totalPages}
+      currentPage={currentPage}
+    />
   );
 }
