+++ src/redux/services/api_people.ts
@@ -1,28 +1,19 @@
-/*import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-
-export const peopleApi = createApi({
-  reducerPath: 'peopleApi',
-  baseQuery: fetchBaseQuery({
-    baseUrl: 'https://swapi.dev/api/people/',
-  }),
-
-  endpoints: (builder) => ({
-    getPeople: builder.query({
-      query: (page) => `?page=${page}`,
-    }),
-    getPeopleId: builder.query({
-      query: (id) => `${id}/`,
-    }),
-    getSearch: builder.query({
-      query: (name) => `?search=${name}`,
-    }),
-  }),
-});
-export const {
-  useGetPeopleQuery,
-  useGetPeopleIdQuery,
-  useGetSearchQuery,
- 
-} = peopleApi;
-// export endpoints for use in SSR
-export const { getPeople, getPeopleId, getSearch } = peopleApi.endpoints;*/
+export async function fetchInvoicesPages(): Promise<number> {
+  const response = await fetch(`https://swapi.dev/api/people`, {
+    cache: 'no-store',
+  });
+  const total = await response.json();
+  return Math.ceil(total.count / 10);
+}
+export async function getPeople(page: number) {
+  const response = await fetch(`https://swapi.dev/api/people?page=${page}`, {
+    cache: 'no-store',
+  });
+  return response.json();
+}
+export async function getSearch(name: string) {
+  const response = await fetch(`https://swapi.dev/api/people?search=${name}`, {
+    cache: 'no-store',
+  });
+  return response.json();
+}
